Enable soundcloud downloads, fix file formats and refactor.
